+++ conman/compat.py
@@ -1,4 +1,0 @@
-import django
-
-
-DJANGO_LT_19 = django.VERSION < (1, 9)