+++ conman/redirects/tests/test_views.py
@@ -1,6 +1,5 @@
 from django.test import TestCase
 
-from conman.compat import DJANGO_LT_19
 from conman.routes.tests.factories import ChildRouteFactory
 from conman.tests.utils import RequestTestCase
 from .factories import ChildRouteRedirectFactory, URLRedirectFactory
@@ -45,10 +44,7 @@ class TestRouteRedirectViewIntegration(TestCase):
         route = ChildRouteRedirectFactory.create(target=target)
         response = self.client.get(route.url)
 
-        expected = target.url
-        if DJANGO_LT_19:
-            expected = 'http://testserver' + expected
-        self.assertEqual(response['Location'], expected)
+        self.assertEqual(response['Location'], target.url)
 
 
 class TestURLRedirectView(RequestTestCase):