+++ conman/routes/handlers.py
@@ -1,4 +1,4 @@
-from django.core.urlresolvers import resolve, Resolver404
+from django.urls import resolve, Resolver404
 
 
 class BaseHandler: