+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # django-conman changelog
 
+## Upcoming release
+
+### Backwards incompatible
+
+* Removed dependancy upon `django-polymorphic-tree`.
+* Destroyed and recreated migrations.
+
 ## 0.0.1a1
 * Unstable, incomplete API
 * Released to claim spot on PyPI