+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 * Removed dependancy upon `django-polymorphic-tree`.
 * Destroyed and recreated migrations.
 
+### Changed
+
+* Moved `RouteManager` into `routes.managers`.
+
 ## 0.0.1a1
 * Unstable, incomplete API
 * Released to claim spot on PyPI